requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
cryptography==41.0.4
PyMySQL==1.1.0
gunicorn==21.2.0
